Create qc_workflow
This workflow downloads Salmonella RNA-seq data for acidic, oxidative, and starvation stress conditions. It performs quality control on the raw reads using FastQC and MultiQC, cleans the reads with fastp, and then re-runs FastQC and MultiQC on the trimmed reads to compare quality before and after cleaning. Download for the Salmonella reference genome and annotation files for downstream analysis are included at the end.

-- Brianna